[improve] sortByIDList move to utils/common
